refactor: parity for struct naming convention

+++ Code/Game/Subsystem/Widget/ButtonWidget.cpp
@@ -55,12 +55,12 @@ String ButtonWidget::GetText() const
 
 void ButtonWidget::SetPosition(Vec2 const& newPosition)
 {
-    m_x = newPosition.x;
-    m_y = newPosition.y;
+    m_x = (int)newPosition.x;
+    m_y = (int)newPosition.y;
 }
 
 void ButtonWidget::SetDimensions(Vec2 const& newDimensions)
 {
-    m_width  = newDimensions.x;
-    m_height = newDimensions.y;
+    m_width  = (int)newDimensions.x;
+    m_height = (int)newDimensions.y;
 }
